refactor(pfb): extract copy/paste into a useClipboard composable
