+++ frappe/public/js/print_format_builder/stores/index.js
@@ -1,24 +1,11 @@
-import { clone_plain, create_default_layout, freshen_field, serialize_layout } from "../utils";
+import { clone_plain, create_default_layout, serialize_layout } from "../utils";
 import { useLayoutHistory } from "./useLayoutHistory";
 import { usePresets } from "../composables/usePresets";
 import { usePreviewDoc } from "../composables/usePreviewDoc";
 import { useSelection } from "../composables/useSelection";
 import { useLayoutMutations } from "../composables/useLayoutMutations";
+import { useClipboard } from "../composables/useClipboard";
 import { watch, ref, inject, computed, nextTick } from "vue";
-
-// Copy/paste clipboard — persisted to localStorage so a field or section copied
-// in one print format can be pasted into another, even across reloads or tabs.
-const CLIPBOARD_KEY = "pfb_clipboard";
-
-function load_clipboard() {
-	try {
-		return JSON.parse(localStorage.getItem(CLIPBOARD_KEY)) || null;
-	} catch {
-		return null;
-	}
-}
-
-const clipboard = ref(load_clipboard());
 
 function persist_local(key, value) {
 	try {
@@ -27,17 +14,6 @@ function persist_local(key, value) {
 	} catch {
 		return false;
 	}
-}
-
-function set_clipboard(value) {
-	clipboard.value = value;
-	persist_local(CLIPBOARD_KEY, value);
-}
-
-if (typeof window !== "undefined") {
-	window.addEventListener("storage", (e) => {
-		if (e.key === CLIPBOARD_KEY) clipboard.value = load_clipboard();
-	});
 }
 
 const SECTION_SNIPPETS_KEY = "pfb_section_snippets";
@@ -283,43 +259,14 @@ export function getStore(print_format_name) {
 		{ deep: true }
 	);
 
-	function copy_field(df) {
-		if (!df) return;
-		set_clipboard({ type: "field", data: clone_plain(df) });
-	}
-	function copy_section(section) {
-		if (!section) return;
-		set_clipboard({ type: "section", data: clone_plain(section) });
-	}
-	function copy_selection() {
-		if (selected_field.value) copy_field(selected_field.value);
-		else if (selected_section.value) copy_section(selected_section.value);
-	}
 	const { style_presets, save_style_preset, apply_style_preset, delete_style_preset } =
 		usePresets(print_format);
-	function paste_clipboard() {
-		const clip = load_clipboard() || clipboard.value;
-		clipboard.value = clip;
-		if (!clip || !layout.value) return;
-
-		if (clip.type === "field") {
-			const clone = freshen_field(clone_plain(clip.data));
-			const col = selected_field.value && find_field_column(selected_field.value);
-			if (col) {
-				col.fields.splice(col.fields.indexOf(selected_field.value) + 1, 0, clone);
-			} else {
-				const sections = layout.value.sections || [];
-				const target = selected_section.value || sections[sections.length - 1];
-				const first_col = target?.columns?.[0];
-				if (!first_col) return;
-				first_col.fields.push(clone);
-			}
-			selected_section.value = null;
-			selected_field.value = clone;
-		} else if (clip.type === "section") {
-			insert_section(clip.data);
-		}
-	}
+	const { clipboard, copy_field, copy_section, copy_selection, paste_clipboard } = useClipboard({
+		selection,
+		layout,
+		find_field_column,
+		insert_section,
+	});
 	const section_snippets = ref(load_section_snippets());
 	function save_section_snippet(name, section) {
 		name = (name || "").trim();
